lambdas/budget-handler/requirements.txt
boto3==1.34.23
botocore==1.34.23
jmespath==1.0.1
mysql-connector-python==8.3.0
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
